Fix delete corpo vvf

+++ api/controllers/corpovvf/delete.js
@@ -11,6 +11,11 @@ module.exports = {
     id: {
       type: 'number',
       required: true
+    },
+    password:{
+      type: 'string',
+      required: true,
+      allowNull: false
     }
   },
 
@@ -21,11 +26,20 @@ module.exports = {
 
 
   fn: async function (inputs) {
+    var corpo = await Corpovvf.findOne({id: inputs.id});
+    if (corpo === null || corpo === undefined || corpo === false){
+      return [{error: '401', message: 'Unauthorized'}];
+    }
+    if (inputs.password !== corpo.password){
+      return [{error: '401', message: 'Unauthorized'}];
+    }
     const record = await Corpovvf.destroy({id: inputs.id}).fetch();
     if (record === 0){
-      throw({invalid: {error: 'Record doesn\'t exist'}});
+      return {error: true};
+      //throw({invalid: {error: 'Record doesn\'t exist'}});
     }
-    return {status: 'OK'};
+    return {error: false};
+    //return {status: 'OK'};
   }
